fix creation and deletion of consultation events, fixes #982
Closes #982

Merge request studip/studip!971

+++ lib/models/ConsultationSlot.php
@@ -193,6 +193,10 @@ class ConsultationSlot extends SimpleORMap
      */
     public function updateEvents()
     {
+        if ($this->isNew()) {
+            return;
+        }
+
         // If no range is associated, remove the event
         if (!$this->block->range) {
             $this->events->delete();
